+++ source/dotnet/wholesale-api/Events/Events.Application/Communication/ICalculationResultIntegrationEventFactory.cs
@@ -12,19 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.Messaging.Communication;
 using Energinet.DataHub.Core.Messaging.Communication.Internal;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model;
 
-namespace Energinet.DataHub.Wholesale.Events.Application.CalculationResultPublishing;
+namespace Energinet.DataHub.Wholesale.Events.Application.Communication;
 
 public interface ICalculationResultIntegrationEventFactory
 {
-    IntegrationEvent CreateForEnergySupplier(CalculationResult result);
-
-    IntegrationEvent CreateForBalanceResponsibleParty(CalculationResult result);
-
-    IntegrationEvent CreateForTotalGridArea(CalculationResult result);
-
-    IntegrationEvent CreateForEnergySupplierByBalanceResponsibleParty(CalculationResult result);
+    IntegrationEvent Create(CalculationResult calculationResult);
 }
