+++ source/dotnet/wholesale-api/Events/Events.Application/UseCases/IUnitOfWork.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Events.Application;
+namespace Energinet.DataHub.Wholesale.Events.Application.UseCases;
 
 public interface IUnitOfWork
 {
