+++ source/dotnet/wholesale-api/Events/Events.Application/UseCases/RegisterCompletedBatchesHandler.cs
@@ -14,7 +14,6 @@
 
 using Energinet.DataHub.Wholesale.Batches.Interfaces;
 using Energinet.DataHub.Wholesale.Events.Application.CompletedBatches;
-using Energinet.DataHub.Wholesale.Events.Application.UseCases.Factories;
 
 namespace Energinet.DataHub.Wholesale.Events.Application.UseCases;
