+++ source/dotnet/wholesale-api/Events/Events.Infrastructure/Persistence/EventsDatabaseContext.cs
@@ -14,7 +14,7 @@
 
 using System.Diagnostics.CodeAnalysis;
 using Energinet.DataHub.Wholesale.Events.Application.CompletedBatches;
-using Energinet.DataHub.Wholesale.Events.Infrastructure.Persistence.Batches;
+using Energinet.DataHub.Wholesale.Events.Infrastructure.Persistence.CompletedBatches;
 using Microsoft.EntityFrameworkCore;
 
 namespace Energinet.DataHub.Wholesale.Events.Infrastructure.Persistence;
