+++ source/dotnet/wholesale-api/Events/Events.Infrastructure/Persistence/UnitOfWork.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Events.Application;
+using Energinet.DataHub.Wholesale.Events.Application.UseCases;
 
 namespace Energinet.DataHub.Wholesale.Events.Infrastructure.Persistence;
