+++ source/dotnet/wholesale-api/Events/Events.UnitTests/Infrastructure/Persistence/Batches/CompletedBatchRepositoryTests.cs
@@ -16,7 +16,7 @@ using AutoFixture.Xunit2;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 using Energinet.DataHub.Wholesale.Events.Application.CompletedBatches;
 using Energinet.DataHub.Wholesale.Events.Infrastructure.Persistence;
-using Energinet.DataHub.Wholesale.Events.Infrastructure.Persistence.Batches;
+using Energinet.DataHub.Wholesale.Events.Infrastructure.Persistence.CompletedBatches;
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.SqlServer.NodaTime.Extensions;
