+++ source/dotnet/wholesale-api/WebApi.IntegrationTests/Fixtures/TestCommon/MessageMetaDataConstants.cs
